+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### Added
 - Added the `favoritesToggleComponent` props to the `Header` component
+- Added an inverse color variant to the `IconButton` component when `inversecolors` is enabled
+- Added the `darkInverse`, `focusInverse`, and `selectedInverse` colors to the palette
 
 ### Fixed
+- Corrected the focus styling of `preview icon` in the `Tile` component and `zoom buttons` in the `Preview` component to meet the required contrast ratio
 
 ### Changed
 - Added keyboard accessibility to the header in `DataGrid` component